+++ cypress/integration/intro.spec.js
@@ -3,7 +3,7 @@ describe('Intro page', () => {
     cy.visit('/');
   });
   it('displays intro text', () => {
-    cy.contains('Here are some examples');
+    cy.contains('Examples');
   });
   it('displays menu', () => {
     cy.contains('Open file');