+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
 
+## 2021-10-31
+
+- Show changelog in intro page and when an upgraded version is loaded
+
 ## 2021-10-27
 
-- Showing events in details panel for WikiTree profiles
+- Show events in details panel for WikiTree profiles
 
 ## 2021-10-26